Call classes for all script functions - database
